backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9
geoalchemy2==0.14.3
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
aiofiles==23.2.1
aiosmtplib==3.0.1
pillow==10.2.0
numpy==1.26.4
scikit-learn==1.4.0
python-dotenv==1.0.1
jinja2==3.1.3
httpx==0.27.0
shapely==2.0.3
email-validator==2.1.0.post1
